+++ wakamiti/wakamiti-api/src/main/java/es/iti/wakamiti/api/WakamitiAPI.java
@@ -5,8 +5,6 @@ import es.iti.wakamiti.api.plan.PlanSerializer;
 import es.iti.wakamiti.api.util.ResourceLoader;
 import imconfig.Configuration;
 
-
-import java.io.File;
 import java.nio.file.Path;
 import java.util.ServiceLoader;
 
@@ -17,24 +15,17 @@ public interface WakamitiAPI {
     }
 
     WakamitiContributors contributors();
+
     ExtensionManager extensionManager();
+
     PlanSerializer planSerializer();
 
-
-    /**
-     * @deprecated Use either {@link #resourceLoader(File)} or {@link #resourceLoader(Configuration)}
-     * instead
-     */
-    @Deprecated
     ResourceLoader resourceLoader();
 
-    ResourceLoader resourceLoader(File workingDir);
-
-    ResourceLoader resourceLoader(Configuration configuration);
-
-
     void publishEvent(String eventType, Object data);
+
     String version();
+
     Path workingDir(Configuration configuration);
 
     Configuration defaultConfiguration();