+++ wakamiti/wakamiti-api/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][1],
 and this project adheres to [Semantic Versioning][2].
 
+## [Unreleased]
+
+## Fixed
+
+- Property `workingDir`
+
 ## [2.1.0] - 2023-07-05
 
 ### Added